+++ pkg/runtime/config/accessors_test.go
@@ -182,6 +182,26 @@ properties:
 		}
 	})
 
+	t.Run("GetStringMapSupportsMapStringStringAndMapAnyAny", func(t *testing.T) {
+		mocks := setupConfigMocks(t)
+		handler := NewConfigHandler(mocks.Shell).(*configHandler)
+
+		handler.data["env_string"] = map[string]string{"A": "a"}
+		envString := handler.GetStringMap("env_string")
+		if envString["A"] != "a" {
+			t.Errorf("Expected map[string]string passthrough, got %v", envString)
+		}
+
+		handler.data["env_any"] = map[any]any{"B": 2, 3: "ignored"}
+		envAny := handler.GetStringMap("env_any")
+		if envAny["B"] != "2" {
+			t.Errorf("Expected map[any]any conversion for string key, got %v", envAny)
+		}
+		if _, exists := envAny["3"]; exists {
+			t.Errorf("Expected non-string key to be ignored, got %v", envAny)
+		}
+	})
+
 	t.Run("SetValidatesPathsAndSchema", func(t *testing.T) {
 		mocks := setupConfigMocks(t)
 		tmpDir, _ := mocks.Shell.GetProjectRoot()
@@ -217,6 +237,71 @@ additionalProperties: false
 
 		if err := handler.Set("disallowed_key", "value"); err == nil {
 			t.Fatal("Expected schema validation error for disallowed key")
+		}
+	})
+}
+
+func TestConfigHandler_AccessorTypeCoercionHelpers(t *testing.T) {
+	t.Run("GetExpectedTypeFromSchemaHandlesPresentAndMissingKeys", func(t *testing.T) {
+		handler, _ := setupPrivateTestHandler(t)
+		handler.schemaValidator.Schema = map[string]any{
+			"properties": map[string]any{
+				"enabled": map[string]any{"type": "boolean"},
+				"count":   map[string]any{"type": "integer"},
+				"ratio":   map[string]any{"type": "number"},
+				"name":    map[string]any{"type": "string"},
+			},
+		}
+
+		if got := handler.getExpectedTypeFromSchema("enabled"); got != "boolean" {
+			t.Errorf("Expected boolean type, got %q", got)
+		}
+		if got := handler.getExpectedTypeFromSchema("missing"); got != "" {
+			t.Errorf("Expected empty type for missing key, got %q", got)
+		}
+
+		handler.schemaValidator.Schema = map[string]any{
+			"properties": map[string]any{"broken": "not-a-map"},
+		}
+		if got := handler.getExpectedTypeFromSchema("broken"); got != "" {
+			t.Errorf("Expected empty type for invalid schema property, got %q", got)
+		}
+	})
+
+	t.Run("ConvertStringToTypeCoversValidAndInvalidConversions", func(t *testing.T) {
+		handler, _ := setupPrivateTestHandler(t)
+
+		if got := handler.convertStringToType("true", "boolean"); got != true {
+			t.Errorf("Expected true boolean, got %v", got)
+		}
+		if got := handler.convertStringToType("42", "integer"); got != 42 {
+			t.Errorf("Expected int 42, got %v", got)
+		}
+		if got := handler.convertStringToType("3.25", "number"); got != 3.25 {
+			t.Errorf("Expected float 3.25, got %v", got)
+		}
+		if got := handler.convertStringToType("abc", "string"); got != "abc" {
+			t.Errorf("Expected string abc, got %v", got)
+		}
+		if got := handler.convertStringToType("not-bool", "boolean"); got != nil {
+			t.Errorf("Expected nil for invalid boolean conversion, got %v", got)
+		}
+		if got := handler.convertStringToType("x", "unknown"); got != nil {
+			t.Errorf("Expected nil for unknown schema type, got %v", got)
+		}
+	})
+
+	t.Run("GetIntHandlesInt64AndUint64Boundaries", func(t *testing.T) {
+		handler, _ := setupPrivateTestHandler(t)
+		maxInt64 := int64(^uint(0) >> 1)
+		handler.data["within"] = maxInt64
+		handler.data["overflow"] = uint64(^uint(0)>>1) + 1
+
+		if got := handler.GetInt("within"); got != int(maxInt64) {
+			t.Errorf("Expected within-bound int64 conversion, got %d", got)
+		}
+		if got := handler.GetInt("overflow", 77); got != 77 {
+			t.Errorf("Expected default on uint64 overflow, got %d", got)
 		}
 	})
 }
